finish animation naive

+++ src/master_CharAnim/CharAnimViewer.h
@@ -27,7 +27,7 @@ protected:
 	void bvhDrawRec(const chara::BVHJoint& bvh, const Transform& f2w, int f);
 
     chara::BVH m_bvh;
-    int m_frameNumber;
+    float m_frameNumber;
 
     Skeleton m_ske;
